# public/openapi.yaml
openapi: 3.1.0
info:
  title: Lune API
  version: 0.3.0
  description: >
    API publique (auth, santé) et API app (PERSONAL / PRO). Toutes les IDs BigInt sont sérialisées en string.
    Toutes les réponses mutables exigent CSRF (Origin) et renvoient `x-request-id` + `cache-control: no-store`.
servers:
  - url: http://localhost:3000
    description: Développement local
  - url: https://diwanbg.work
    description: Production (Railway)
security:
  - cookieAuth: []
...
paths:
...
  /api/pro/businesses/{businessId}/processes/{processId}:
    get:
      tags: [Pro / Processes]
      security: [ { cookieAuth: [] } ]
      summary: Détail process (avec étapes)
      parameters:
        - $ref: '#/components/parameters/BusinessId'
        - $ref: '#/components/parameters/ProcessId'
      responses:
        '200':
          description: Process + étapes
          content:
            application/json:
              schema:
                type: object
                properties:
                  item: { $ref: '#/components/schemas/ProcessWithSteps' }
    patch:
      tags: [Pro / Processes]
      security: [ { cookieAuth: [] } ]
      summary: Met à jour un process (nom, description, archive/unarchive)
      description: ADMIN/OWNER. CSRF (Origin) requis + rate limit pro:processes:update.
      parameters:
        - $ref: '#/components/parameters/BusinessId'
        - $ref: '#/components/parameters/ProcessId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProcessUpdate' }
      responses:
        '200':
          description: Process mis à jour
          content:
            application/json:
              schema:
                type: object
                properties:
                  item: { $ref: '#/components/schemas/ProcessWithSteps' }
    delete:
      tags: [Pro / Processes]
      security: [ { cookieAuth: [] } ]
      summary: Supprime un process
      description: ADMIN/OWNER. CSRF (Origin) requis + rate limit pro:processes:delete.
      parameters:
        - $ref: '#/components/parameters/BusinessId'
        - $ref: '#/components/parameters/ProcessId'
      responses:
        '204':
          description: Process supprimé
  /api/pro/businesses/{businessId}/processes/{processId}/steps:
    post:
      tags: [Pro / Processes]
      security: [ { cookieAuth: [] } ]
      summary: Ajoute une étape
      description: ADMIN/OWNER. CSRF (Origin) requis + rate limit pro:processes:steps:create.
      parameters:
        - $ref: '#/components/parameters/BusinessId'
        - $ref: '#/components/parameters/ProcessId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProcessStepCreate' }
      responses:
        '201':
          description: Étape créée
          content:
            application/json:
              schema:
                type: object
                properties:
                  item: { $ref: '#/components/schemas/ProcessStep' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  parameters:
    BusinessId:
      name: businessId
      in: path
      required: true
      schema: { type: string }
...
    ProcessId:
      name: processId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    ProcessStatus:
      type: string
      enum: [ACTIVE, ARCHIVED]
    Process:
      type: object
      properties:
        id: { type: string }
        businessId: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/ProcessStatus' }
        stepsCount: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProcessWithSteps:
      allOf:
        - $ref: '#/components/schemas/Process'
        - type: object
          properties:
            steps:
              type: array
              items: { $ref: '#/components/schemas/ProcessStep' }
...
    ProcessUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/ProcessStatus' }
        archived: { type: boolean, nullable: true }
    ProcessStep:
      type: object
      properties:
        id: { type: string }
        processId: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        position: { type: integer }
        isDone: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProcessStepCreate:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
        position: { type: integer, nullable: true }
        isDone: { type: boolean, nullable: true }
...